requirements.txt
python-pptx
pandas
pyarrow